update - criteria plus exercise distance

+++ backend/src/main/java/org/ssafy/d210/walk/dto/request/CustomExerciseCriteriaRequestDto.java
@@ -7,5 +7,5 @@ public class CustomExerciseCriteriaRequestDto {
     private Long steps;
     private Long exerciseMinute;
     private Long heartRate;
-    private Long exerciseDistance;
+    private Double exerciseDistance;
 }
