+++ backend/src/main/java/org/ssafy/d210/walk/dto/response/MainCriteriaResponseDto.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class MainCriteriaResponseDto {
     private Long steps;
     private Long exerciseMinute;
-    private Long exerciseDistance;
+    private Double exerciseDistance;
 }
